pyproject.toml
[project]
name = "gen-ai-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.9.0",
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "groq>=0.31.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.0",
    "openai>=1.107.3",
    "praw>=7.8.1",
    "python-dotenv>=1.1.1",
    "requests>=2.25.0",
    "streamlit>=1.49.1",
    "wikipedia>=1.4.0",
]
